+++ packages/cipherstash-proxy-integration/src/update/mod.rs
@@ -3,3 +3,4 @@ mod update_with_literal;
 mod update_with_null_literal;
 mod update_with_null_param;
 mod update_with_param;
+mod update_with_reused_param;